Working with NewApplication form

+++ app/src/main/java/com/s/samsungitschool/recom00/fragments/NewAppFragmentActivity.java
@@ -277,8 +277,8 @@ public class NewAppFragmentActivity extends Fragment {
                         loadSuccessfully = false;
 
                     } else {
-                        pointId = Long.valueOf(addPointServerAns);
-                        
+                        pointId = (long) Long.valueOf(addPointServerAns);
+
                         loadSuccessfully = true;
                     }
                 }
